Compiler warning fixes

+++ Oxygen/oxygenengine/source/oxygen_netcore/network/internal/CryptoFunctions.cpp
@@ -63,7 +63,7 @@ namespace sha1
 	void SHA1::update(std::istream &is)
 	{
 		std::string rest_of_buffer;
-		read(is, rest_of_buffer, BLOCK_BYTES - buffer.size());
+		read(is, rest_of_buffer, (int)BLOCK_BYTES - (int)buffer.size());
 		buffer += rest_of_buffer;
 
 		while (is)
@@ -82,7 +82,7 @@ namespace sha1
 
 		/* Padding */
 		buffer += (char)0x80;
-		unsigned int orig_size = buffer.size();
+		unsigned int orig_size = (unsigned int)buffer.size();
 		while (buffer.size() < BLOCK_BYTES)
 		{
 			buffer += (char)0x00;
